Adding math function calls and some fixes

+++ lib/api/RainViewer.sh
@@ -25,35 +25,44 @@ generate_radar_pic() {
     local response="$5"
 
     local crop_size crop_L crop_T
-    read -r crop_size crop_L crop_T <<<"$(zoom_calculi "$virtual_zoom")"
+    read -r crop_size crop_L crop_T <<<"$(awk -f lib/math.awk -v zoom="$virtual_zoom" -v pic_size="$PICTURE_SIZE" -e 'BEGIN { print zoom_calculi(zoom, pic_size ) }')"
 
     local image_path
     image_path="$(jq -r '.host + .radar.past[-1].path' <<<"$response")"
+
+    local radar_pic
+    radar_pic="$CACHE_DIR/$(jq -r '.radar.past.[-1].time' <<<"$response")_${c_lat}_${c_lon}_${api_zoom}.png"
 
     local tmpdir
     tmpdir=$(mktemp -d)
 
     local tile_x tile_y offset_L offset_T
-    read -r tile_x tile_y offset_L offset_T <<<"$(latlon_to_tile "$c_lat" "$c_lon" "$api_zoom")"
+    read -r tile_x tile_y offset_L offset_T <<<"$(awk -f lib/math.awk -v lat="$c_lat" -v lon="$c_lon" -v zoom="$api_zoom" -e 'BEGIN { print latlon_to_tile(lat, lon, zoom) }')"
 
     local i=0 tx ty
-    for y in -1 0 1; do
-        for x in -1 0 1; do
-            tx=$((tile_x + x))
-            ty=$((tile_y + y))
-            curl -sf "$image_path/$((PICTURE_SIZE / 2))/$api_zoom/$tx/$ty/1/1_1.png" -o "$tmpdir/tile_${i}.png" &
-            ((i++))
+    if [[ ! -f "$radar_pic" ]]; then
+        for y in -1 0 1; do
+            for x in -1 0 1; do
+                tx=$((tile_x + x))
+                ty=$((tile_y + y))
+                curl -sf "$image_path/$((PICTURE_SIZE / 2))/$api_zoom/$tx/$ty/1/1_1.png" -o "$tmpdir/tile_${i}.png" &
+                ((i++))
+            done
         done
-    done
-    wait
+        wait
 
-    magick "$tmpdir/tile_0.png" "$tmpdir/tile_1.png" "$tmpdir/tile_2.png" +append \
-        \( "$tmpdir/tile_3.png" "$tmpdir/tile_4.png" "$tmpdir/tile_5.png" +append \) \
-        \( "$tmpdir/tile_6.png" "$tmpdir/tile_7.png" "$tmpdir/tile_8.png" +append \) \
-        -append \
-        -crop "${PICTURE_SIZE}x${PICTURE_SIZE}+${offset_L}+${offset_T}" +repage \
-        -filter point \
-        "$RAINVIEWER_PIC" 2>/dev/null
+        magick "$tmpdir/tile_0.png" "$tmpdir/tile_1.png" "$tmpdir/tile_2.png" +append \
+            \( "$tmpdir/tile_3.png" "$tmpdir/tile_4.png" "$tmpdir/tile_5.png" +append \) \
+            \( "$tmpdir/tile_6.png" "$tmpdir/tile_7.png" "$tmpdir/tile_8.png" +append \) \
+            -append \
+            -crop "${PICTURE_SIZE}x${PICTURE_SIZE}+${offset_L}+${offset_T}" +repage \
+            -filter point \
+            "$RAINVIEWER_PIC" 2>/dev/null
+        cp "$RAINVIEWER_PIC" "$radar_pic"
+        rm -rf "$tmpdir"
+    else
+        magick "$radar_pic" "$RAINVIEWER_PIC" 2>/dev/null
+    fi
 
     magick "$RAINVIEWER_PIC" \
         -crop "${crop_size}x${crop_size}+${crop_L}+${crop_T}" +repage \
@@ -61,5 +70,4 @@ generate_radar_pic() {
         -resize "${PICTURE_SIZE}x${PICTURE_SIZE}" \
         "$CROPPED_RADAR_PIC" 2>/dev/null
 
-    rm -rf "$tmpdir"
 }
